refactor ray.color functionality to integrator
