+++ src/tracer/object/plane_tests.rs
@@ -13,7 +13,6 @@ fn no_self_intersect() {
     let r = Ray::new(
         DVec3::ZERO,
         DVec3::ONE,
-        0,
     );
     assert!(plane().hit(&r).is_none());
 }
@@ -23,7 +22,6 @@ fn no_intersect_behind() {
     let r = Ray::new(
         DVec3::ONE,
         DVec3::ONE,
-        0,
     );
     assert!(plane().hit(&r).is_none());
 }
@@ -33,7 +31,6 @@ fn intersects() {
     let r = Ray::new(
         -DVec3::ONE,
         DVec3::ONE,
-        0,
     );
     assert!(plane().hit(&r).is_some());
 }
@@ -45,7 +42,6 @@ fn no_hit_crash_parallel() {
         -DVec3::ONE,
         // pray it works
         p.norm.cross(DVec3::new(1.23, 4.56, 7.89)),
-        0,
     );
     assert!(p.hit(&r).is_none());
 }
